ページのパスの統一

+++ src/database/migrations/2024_06_30_105810_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('product_id');
             $table->string('status')->default('pending');
-            $table->integer('amount');
+            $table->unsignedBigInteger('amount');
             $table->string('payment_method');
             $table->timestamps();
         });
